+++ packages/localizations/src/es-ES.ts
@@ -1761,6 +1761,7 @@ export const esES: LocalizationResource = {
     form_password_compromised__sign_in: undefined,
     form_password_incorrect: 'Contraseña incorrecta.',
     form_password_length_too_short: 'La contraseña es demasiado corta.',
+    form_password_matches_identifier: undefined,
     form_password_not_strong_enough: 'Tu contraseña no es lo suficientemente fuerte.',
     form_password_or_identifier_incorrect:
       'La contraseña o la dirección de correo electrónico es incorrecta. Inténtalo de nuevo o usa otro método.',