+++ packages/tanstack-react-start/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clerk/tanstack-react-start
 
+## 1.5.2
+
+### Patch Changes
+
+- Updated dependencies [[`7f5c294`](https://github.com/clerk/javascript/commit/7f5c2947e2b3b2ac9677116ff7eede61a1dab649)]:
+  - @clerk/shared@4.29.1
+  - @clerk/backend@3.16.6
+  - @clerk/react@6.14.3
+
 ## 1.5.1
 
 ### Patch Changes